feat: wrap all viz components in QueryLoad
