+++ packages/core-components/src/lib/atoms/index.js
@@ -4,4 +4,4 @@ export * from './skeletons';
 export * from './dropdown';
 export * from './button';
 export { default as CopyButton } from './copy-button/CopyButton.svelte';
-export * from "./query-load";
+export * from './query-load';
