+++ packages/core-components/src/lib/atoms/query-load/index.js
@@ -1,1 +1,1 @@
-export {default as QueryLoad} from "./QueryLoad.svelte"
+export { default as QueryLoad } from './QueryLoad.svelte';
