+++ packages/core-components/src/lib/unsorted/QueryStatus.svelte
@@ -55,12 +55,13 @@
 			dismissable: true
 		};
 
-		toasts.add(status_as_toast, 0);
+		toasts.add(status_as_toast, 5000);
 	};
 
 	onMount(() => {
 		if (!browser) return;
 		if (!import.meta.hot) return;
+		import.meta.hot.off?.('evidence:build-status', handleStatusEvent);
 		import.meta.hot.on?.('evidence:build-status', handleStatusEvent);
 	});
