+++ packages/core-components/src/lib/unsorted/viz/core/Value.svelte
@@ -3,18 +3,18 @@
 </script>
 
 <script>
-	import {QueryLoad} from "../../../atoms/query-load"
-	import Value from "./_Value.svelte"
-	export let data
+	import { QueryLoad } from '../../../atoms/query-load';
+	import Value from './_Value.svelte';
+	export let data;
 
 	// Remove any undefined props (e.g. w/o defaults) to prevent them from being passed
-	$: spreadProps = Object.fromEntries(Object.entries($$props).filter(([k,v]) => v !== undefined))
+	$: spreadProps = Object.fromEntries(Object.entries($$props).filter(([, v]) => v !== undefined));
 </script>
 
 <!-- Pass all the props through-->
 <QueryLoad {data}>
-    <p slot="skeleton">Loading...</p>
+	<p slot="skeleton">Loading...</p>
 	<Value {...spreadProps} data={data ? Array.from(data) : null}>
-		<slot/>
+		<slot />
 	</Value>
 </QueryLoad>
