+++ packages/core-components/src/lib/unsorted/viz/references/ReferenceArea.svelte
@@ -228,7 +228,6 @@
 			}
 			return d;
 		});
-		console.log($config);
 	});
 </script>
