+++ sites/test-env/sources/social_media/posts/post_tags.yaml
@@ -1,4 +1,4 @@
-rows: 6500
+rows: 65000
 fuzz: 1000
 schema:
     id:
