+++ sites/test-env/sources/social_media/posts/posts.yaml
@@ -1,4 +1,4 @@
-rows: 25000
+rows: 250000
 fuzz: 2500
 schema:
     id:
