+++ sites/test-env/svelte.config.js
@@ -1,6 +1,9 @@
 /** @type {import("@sveltejs/kit").Config} */
 export default {
 	preprocess: [],
+	compilerOptions: {
+		dev: true
+	},
 	kit: {},
 	vite: {
 		server: {
